web-app/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymysql==1.1.0
PyJWT==2.8.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.2
onnxruntime==1.16.3
gunicorn==21.2.0
